+++ hw-01/src/main/java/ru/otus/dao/QuestionDao.java
@@ -1,0 +1,10 @@
+package ru.otus.dao;
+
+import ru.otus.domain.Question;
+
+import java.util.List;
+
+public interface QuestionDao {
+
+    List<Question> getQuestions();
+}
