--- hw-01/src/main/java/ru/otus/service/PrintService.java
@@ -1,10 +1,0 @@
-package ru.otus.service;
-
-import ru.otus.domain.Question;
-
-import java.util.List;
-
-public interface PrintService {
-
-    void printQuestions(List<Question> questionList);
-}
