--- hw-01/src/main/java/ru/otus/service/QuestionLoaderService.java
@@ -1,10 +1,0 @@
-package ru.otus.service;
-
-import ru.otus.domain.Question;
-
-import java.util.List;
-
-public interface QuestionLoaderService {
-
-    List<Question> loadQuestions();
-}
