+++ hw-01/src/main/java/ru/otus/service/QuestionService.java
@@ -1,0 +1,6 @@
+package ru.otus.service;
+
+public interface QuestionService {
+
+    void printQuestions();
+}
