+++ hw-01/src/main/java/ru/otus/service/QuestionServiceImpl.java
@@ -1,12 +1,18 @@
 package ru.otus.service;
 
+import lombok.RequiredArgsConstructor;
+import ru.otus.dao.QuestionDao;
 import ru.otus.domain.Question;
 
 import java.util.List;
 
-public class ConsolePrintService implements PrintService {
+@RequiredArgsConstructor
+public class QuestionServiceImpl implements QuestionService {
+    private final QuestionDao questionDao;
+
     @Override
-    public void printQuestions(List<Question> questionList) {
+    public void printQuestions() {
+        List<Question> questionList = questionDao.getQuestions();
         for (var question : questionList) {
             System.out.println(question.getValue());
             for (var answer : question.getAnswers()) {
