+++ hw-01/src/main/resources/spring-context.xml
@@ -2,8 +2,10 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-    <bean id="questionLoaderService" class="ru.otus.service.CsvQuestionLoaderService">
+    <bean id="questionDao" class="ru.otus.dao.CsvQuestionDao">
         <constructor-arg index="0" value="/questions.csv"/>
     </bean>
-    <bean id="printService" class="ru.otus.service.ConsolePrintService"/>
+    <bean id="questionService" class="ru.otus.service.QuestionServiceImpl">
+        <constructor-arg index="0" ref="questionDao"/>
+    </bean>
 </beans>
